Move Stream class into diff module
