+++ pipeline/cloud/schemas/runs.py
@@ -5,6 +5,7 @@ from datetime import datetime
 from enum import Enum
 
 from pipeline.cloud.schemas import BaseModel
+from pipeline.objects.variables import Stream
 
 
 class RunState(str, Enum):
@@ -132,9 +133,7 @@ class RunIOType(str, Enum):
                 return cls.pkl
             return cls.array
         # Also check if class name == Stream
-        elif hasattr(obj, "iterable") or (
-            hasattr(obj, "__name__") and obj.__name__ == "Stream"
-        ):
+        elif isinstance(obj, Stream):
             return cls.stream
         elif (
             isinstance(obj, io.BufferedIOBase)
